moved csv logic from router and backend-processing to ParsingController
